refactor(card): simplify card token to use two uuids without dash
