+++ card/src/core/card/repository.go
@@ -1,0 +1,6 @@
+package card
+
+type CardRepository interface {
+	Generate(generateCardDTO *GenerateCardRepoInput) (*Card, error)
+	GetByToken(token string) (*Card, error)
+}
