--- card/src/core/encrypter/encrypter.go
@@ -1,73 +1,0 @@
-package encrypter
-
-import (
-	"bytes"
-	"crypto/aes"
-	"crypto/cipher"
-	b64 "encoding/base64"
-	"os"
-)
-
-type Encrypter interface {
-	Encrypt(content []byte) (encryptedContent []byte, err error)
-	Decrypt(encryptedData []byte) (decryptedContent []byte, err error)
-}
-
-type encryter struct {
-	encryptionKey []byte
-}
-
-func NewEncrypter(encryptionKey []byte) *encryter {
-	return &encryter{
-		encryptionKey: encryptionKey,
-	}
-}
-
-func (e *encryter) Encrypt(content []byte) (encryptedContent []byte, err error) {
-	plainText := pKCS5Padding(content, aes.BlockSize)
-	ciphertext := make([]byte, len(plainText))
-
-	block, err := aes.NewCipher(e.encryptionKey)
-
-	if err != nil {
-		return nil, err
-	}
-
-	IV := []byte(os.Getenv("ENCRYPTION_IV"))
-	mode := cipher.NewCBCEncrypter(block, IV)
-	mode.CryptBlocks(ciphertext, plainText)
-	encoded := b64.RawURLEncoding.EncodeToString(ciphertext)
-	return []byte(encoded), nil
-}
-
-func (e *encryter) Decrypt(encryptedData []byte) (decryptedContent []byte, err error) {
-	iv := []byte(os.Getenv("ENCRYPTION_IV"))
-	decodedData, err := b64.RawURLEncoding.DecodeString(string(encryptedData))
-
-	if err != nil {
-		return nil, err
-	}
-
-	block, err := aes.NewCipher(e.encryptionKey)
-	if err != nil {
-		return nil, err
-	}
-
-	mode := cipher.NewCBCDecrypter(block, iv)
-	mode.CryptBlocks(decodedData, decodedData)
-
-	unpaddedData := pKCS5UnPadding(decodedData)
-	return unpaddedData, nil
-}
-
-func pKCS5Padding(cipherText []byte, blockSize int) []byte {
-	padding := blockSize - len(cipherText)%blockSize
-	padText := bytes.Repeat([]byte{byte(padding)}, padding)
-	return append(cipherText, padText...)
-}
-
-func pKCS5UnPadding(src []byte) []byte {
-	length := len(src)
-	unpadding := int(src[length-1])
-	return src[:(length - unpadding)]
-}
