+++ card/src/grpc/grpc.go
@@ -7,16 +7,14 @@ import (
 	postgrescardrepository "github.com/bmviniciuss/tcc/card/src/adapter/card"
 	carddetailsgenerator "github.com/bmviniciuss/tcc/card/src/adapter/carddetails"
 	"github.com/bmviniciuss/tcc/card/src/core/card"
-	"github.com/bmviniciuss/tcc/card/src/core/encrypter"
 	"github.com/bmviniciuss/tcc/card/src/grpc/pb"
 	"github.com/jmoiron/sqlx"
 )
 
 func newCardService(db *sqlx.DB) *card.CardService {
 	cardRepository := postgrescardrepository.NewPostgresCardRepository(db)
-	encrypter := encrypter.NewEncrypter([]byte("gFvJR96@UXYrq_2m"))
 	cardDetailsGenerator := carddetailsgenerator.NewCardDetailsGenerator()
-	cardService := card.NewCardService(cardDetailsGenerator, encrypter, cardRepository)
+	cardService := card.NewCardService(cardDetailsGenerator, cardRepository)
 	return cardService
 }
